@@ -11,25 +11,31 @@
       "role": "Developer"
     }
   ],
+  "support": {
+    "issues": "https://github.com/thefrosty/wp-upgrade-task-runner/issues"
+  },
   "config": {
     "sort-packages": true,
-    "optimize-autoloader": true
+    "preferred-install": "dist",
+    "optimize-autoloader": true,
+    "vendor-dir": "vendor"
   },
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": ">=7.1",
+    "php": ">=7.3",
     "pimple/pimple": "^3.2",
-    "symfony/http-foundation": "^3.4||^4",
+    "symfony/http-foundation": "^4.2.12||^5",
     "thefrosty/wp-utilities": "^1.7"
   },
   "require-dev": {
     "dealerdirect/phpcodesniffer-composer-installer": "^0.4.3",
-    "inpsyde/php-coding-standards": "^0.13.3",
-    "phpmd/phpmd": "^2.6",
-    "phpunit/phpunit": "^6.5",
+    "phpcompatibility/php-compatibility": "^9.3",
+    "phpunit/phpunit": "^7",
+    "slevomat/coding-standard": "^4",
     "squizlabs/php_codesniffer": "^3.2",
-    "wp-phpunit/wp-phpunit": "^4"
+    "wp-coding-standards/wpcs": "^2.2",
+    "wp-phpunit/wp-phpunit": "^5.1"
   },
   "autoload": {
     "psr-4": {
@@ -51,15 +57,11 @@
     "phpcs": [
       "bash ./bin/phpcs.sh"
     ],
-    "phpmd": [
-      "bash ./bin/phpmd.sh"
-    ],
     "phpunit": [
       "phpunit --colors"
     ],
     "tests": [
       "@phpcs",
-      "@phpmd",
       "@phpunit"
     ]
   }
