@@ -8,19 +8,21 @@ use TheFrosty\WpUpgradeTaskRunner\Models\UpgradeModel;
  * Interface TaskRunnerInterface
  *
  * @package TheFrosty\WpUpgradeTaskRunner\Api
+ * phpcs:disable SlevomatCodingStandard.Classes.SuperfluousAbstractClassNaming.SuperfluousPrefix
  */
 interface TaskRunnerInterface
 {
-    const DATE = null;
-    const DESCRIPTION = null;
-    const TITLE = null;
+
+    public const DATE = null;
+    public const DESCRIPTION = null;
+    public const TITLE = null;
 
     /**
      * Dispatch the migration task.
      *
      * @param UpgradeModel $model
      */
-    public function dispatch(UpgradeModel $model);
+    public function dispatch(UpgradeModel $model): void;
 
     /**
      * Trigger events when the task is complete.
@@ -34,14 +36,19 @@ interface TaskRunnerInterface
      * Schedule a one off cron event.
      *
      * @param string $class The fully-qualified class-name to register as a cron hook.
-     * @param array $args Optional. Arguments to pass to the hook's callback function.
+     * @param UpgradeModel $model Arguments to pass to the hook's callback function.
+     * @uses wp_schedule_single_event()
      */
-    public function scheduleEvent(string $class, array $args = []);
+    public function scheduleEvent(string $class, UpgradeModel $model): void;
 
     /**
      * Clear the one off scheduled cron event, in-case it isn't cleared automatically.
      *
      * @param string $class The fully-qualified class-name to register as a cron hook.
+     * @param UpgradeModel $model Arguments to pass to the hook's callback function.
+     * @uses wp_clear_scheduled_hook()
+     * @uses wp_next_scheduled()
+     * @uses wp_unschedule_event()
      */
-    public function clearScheduledEvent(string $class);
+    public function clearScheduledEvent(string $class, UpgradeModel $model): void;
 }
