@@ -5,7 +5,6 @@ namespace TheFrosty\WpUpgradeTaskRunner;
 use Pimple\Container;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use TheFrosty\WpUpgradeTaskRunner\Models\UpgradeModel;
-use TheFrosty\WpUpgradeTaskRunner\Tasks\TaskLoader;
 use TheFrosty\WpUtilities\Plugin\AbstractHookProvider;
 use TheFrosty\WpUtilities\Plugin\HttpFoundationRequestInterface;
 use TheFrosty\WpUtilities\Plugin\HttpFoundationRequestTrait;
@@ -16,6 +15,7 @@ use TheFrosty\WpUtilities\Plugin\HttpFoundationRequestTrait;
  */
 class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInterface
 {
+
     use HttpFoundationRequestTrait;
 
     public const AJAX_ACTION = 'wp_upgrade_schedule_task_event';
@@ -24,6 +24,7 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
     public const NONCE_NAME = '_task_runner_execute_nonce';
     public const OPTION_NAME = 'wp_upgrade_task_runner';
     private const NONCE_KEY = 'task_runner_migration_upgrades_nonce_%s';
+    public const TAG_SETTINGS_PAGE_LOADED = 'wp_upgrade_task_runner/settings_page_loaded';
 
     /**
      * Container object.
@@ -45,7 +46,7 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
 
     /**
      * TaskLoader object.
-     * @var TaskLoader $task_loader
+     * @var Tasks\TaskLoader $task_loader
      */
     private $task_loader;
 
@@ -62,26 +63,11 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
     /**
      * Add class hooks.
      */
-    public function addHooks()
+    public function addHooks(): void
     {
         $this->addAction('admin_menu', [$this, 'addDashboardPage']);
         $this->addAction('admin_enqueue_scripts', [$this, 'enqueueScripts']);
         $this->addAction('wp_ajax_' . self::AJAX_ACTION, [$this, 'scheduleTaskRunnerEvent']);
-    }
-
-    /**
-     * Get the options.
-     * @return array
-     */
-    public function getOptions(): array
-    {
-        static $options;
-
-        if (empty($options)) {
-            $options = \get_option(Upgrade::OPTION_NAME, []);
-        }
-
-        return $options;
     }
 
     /**
@@ -98,10 +84,10 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
      * Register our dashboard page.
      * `add_dashboard_page()` returns false if the current user doesn't have the capability.
      */
-    protected function addDashboardPage()
+    protected function addDashboardPage(): void
     {
         if (!\class_exists('WP_List_Table')) {
-            require_once ABSPATH . 'wp-admin/includes/class-wp-list-table.php';
+            require_once \ABSPATH . 'wp-admin/includes/class-wp-list-table.php';
         }
         $this->list_table = new UpgradesListTable($this->container);
         $this->settings_page = \add_dashboard_page(
@@ -109,14 +95,16 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
             \sprintf(\__('Migration Tasks %s', 'wp-upgrade-task-runner'), $this->getUpgradeCountHtml()),
             'promote_users',
             self::MENU_SLUG,
-            function () {
+            function (): void {
                 include $this->getPlugin()->getPath('/views/wp-admin/settings-pages/upgrades.php');
             }
         );
-        if (\is_string($this->settings_page)) {
-            $this->task_loader->registerScreenId($this->settings_page);
-            $this->addAction('load-' . $this->settings_page, [$this, 'maybeExecute']);
+        if (!\is_string($this->settings_page)) {
+            return;
         }
+
+        $this->task_loader->registerScreenId($this->settings_page);
+        $this->addAction('load-' . $this->settings_page, [$this, 'maybeExecute']);
     }
 
     /**
@@ -124,7 +112,7 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
      * page when a tasks description length is too long.
      * @param string $hook_suffix Admin page suffix.
      */
-    protected function enqueueScripts(string $hook_suffix)
+    protected function enqueueScripts(string $hook_suffix): void
     {
         if ($this->settings_page !== $hook_suffix) {
             return;
@@ -181,11 +169,10 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
     /**
      * On our settings page, check for our item and a valid nonce key so we
      * can execute a single task to our cron.
-     * @SuppressWarnings(PHPMD.ExitExpression)
-     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
      */
-    protected function maybeExecute()
+    protected function maybeExecute(): void
     {
+        \do_action(Upgrade::TAG_SETTINGS_PAGE_LOADED, $this->task_loader->getFields(), $this->getRequest());
         $query = $this->getRequest()->query;
         if ((!$query->has(self::NONCE_NAME) || !$query->has('item')) &&
             ($query->get(self::NONCE_NAME) !== null || $query->get('item') !== null)
@@ -209,7 +196,7 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
     /**
      * AJAX listener to trigger a scheduled event.
      */
-    protected function scheduleTaskRunnerEvent()
+    protected function scheduleTaskRunnerEvent(): void
     {
         try {
             $fields = (new \ReflectionClass($this->task_loader))->getMethod('registerFields');
@@ -255,7 +242,7 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
     {
         $key = \array_search(
             \rawurldecode($bag->get('item')),
-            \array_map(function (UpgradeModel $model): string {
+            \array_map(static function (UpgradeModel $model): string {
                 return $model->getTitle();
             }, $this->task_loader->getFields()),
             true
@@ -267,11 +254,10 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
              * @var UpgradeModel $model
              */
             $model = $this->task_loader->getFields()[$key];
-            $options = $this->getOptions();
-            if (!isset($options[\sanitize_title($model->getTitle())]) ||
-                empty($options[\sanitize_title($model->getTitle())])
-            ) {
-                $model->getTaskRunner()->scheduleEvent(\get_class($model->getTaskRunner()), [$model]);
+            $options = Option::getOptions();
+            $option_key = Option::getOptionKey($model);
+            if (empty($options[$option_key])) {
+                $model->getTaskRunner()->scheduleEvent(\get_class($model->getTaskRunner()), $model);
                 return true;
             }
         }
@@ -302,6 +288,6 @@ class Upgrade extends AbstractHookProvider implements HttpFoundationRequestInter
      */
     private function getUpgradeCount(): int
     {
-        return \absint(\count($this->task_loader->getFields()) - \count($this->getOptions()));
+        return \absint(\count($this->task_loader->getFields()) - \count(Option::getOptions()));
     }
 }
