@@ -6,12 +6,12 @@
  * Plugin Name: Upgrade Task Runner
  * Plugin URI: https://github.com/thefrosty/wp-upgrade-task-runner
  * Description: A WordPress plugin for developers to write custom migration tasks.
- * Version: 1.3.0
+ * Version: 2.0.0
  * Author: Austin Passy
  * Author URI: https://github.com/thefrosty
- * Requires at least: 4.8
+ * Requires at least: 5.1
  * Tested up to: 5.2
- * Requires PHP: 7.1
+ * Requires PHP: 7.3
  *
  * @package TheFrosty\WpUpgradeTaskRunner
  */
@@ -19,8 +19,9 @@
 namespace TheFrosty\WpUpgradeTaskRunner;
 
 const SLUG = 'wp-upgrade-task-runner';
-const VERSION = '1.3.0';
+const VERSION = '2.0.0';
 
+use TheFrosty\WpUpgradeTaskRunner\Upgrade\DbUpgrade;
 use TheFrosty\WpUtilities\Plugin\Container;
 use TheFrosty\WpUtilities\Plugin\PluginFactory;
 use TheFrosty\WpUtilities\WpAdmin\DisablePluginUpdateCheck;
@@ -32,8 +33,9 @@ $container->register(new ServiceProvider());
 $plugin
     ->add(new DisablePluginUpdateCheck())
     ->add($container[ServiceProvider::UPGRADE_PROVIDER])
-    ->add($container[ServiceProvider::TASK_LOADER]);
+    ->add($container[ServiceProvider::TASK_LOADER])
+    ->addOnHook(DbUpgrade::class, 'admin_menu', null, true);
 
-\add_action('plugins_loaded', static function () use ($plugin) {
+\add_action('plugins_loaded', static function () use ($plugin): void {
     $plugin->initialize();
 }, 5);
